+++ test/support/sample_live/some_component.ex
@@ -1,0 +1,4 @@
+defmodule SampleLive.SomeComponent do
+  use Phoenix.LiveComponent
+  use Phoenix.LiveController.ViewRenderer
+end